Moved venues booking and non sulten related items out of sulten app
