+++ backend/root/utils/routes.py
@@ -206,14 +206,14 @@ admin__samfundet_blogpost_history = 'admin:samfundet_blogpost_history'
 admin__samfundet_blogpost_delete = 'admin:samfundet_blogpost_delete'
 admin__samfundet_blogpost_change = 'admin:samfundet_blogpost_change'
 adminsamfundetblogpost__objectId = ''
-admin__samfundet_table_permissions = 'admin:sulten_table_permissions'
-admin__samfundet_table_permissions_manage_user = 'admin:sulten_table_permissions_manage_user'
-admin__samfundet_table_permissions_manage_group = 'admin:sulten_table_permissions_manage_group'
-admin__samfundet_table_changelist = 'admin:sulten_table_changelist'
-admin__samfundet_table_add = 'admin:sulten_table_add'
-admin__samfundet_table_history = 'admin:sulten_table_history'
-admin__samfundet_table_delete = 'admin:sulten_table_delete'
-admin__samfundet_table_change = 'admin:sulten_table_change'
+admin__samfundet_table_permissions = 'admin:samfundet_table_permissions'
+admin__samfundet_table_permissions_manage_user = 'admin:samfundet_table_permissions_manage_user'
+admin__samfundet_table_permissions_manage_group = 'admin:samfundet_table_permissions_manage_group'
+admin__samfundet_table_changelist = 'admin:samfundet_table_changelist'
+admin__samfundet_table_add = 'admin:samfundet_table_add'
+admin__samfundet_table_history = 'admin:samfundet_table_history'
+admin__samfundet_table_delete = 'admin:samfundet_table_delete'
+admin__samfundet_table_change = 'admin:samfundet_table_change'
 adminsamfundettable__objectId = ''
 admin__samfundet_reservation_permissions = 'admin:sulten_reservation_permissions'
 admin__samfundet_reservation_permissions_manage_user = 'admin:sulten_reservation_permissions_manage_user'
@@ -269,14 +269,14 @@ admin__samfundet_saksdokument_history = 'admin:samfundet_saksdokument_history'
 admin__samfundet_saksdokument_delete = 'admin:samfundet_saksdokument_delete'
 admin__samfundet_saksdokument_change = 'admin:samfundet_saksdokument_change'
 adminsamfundetsaksdokument__objectId = ''
-admin__samfundet_booking_permissions = 'admin:sulten_booking_permissions'
-admin__samfundet_booking_permissions_manage_user = 'admin:sulten_booking_permissions_manage_user'
-admin__samfundet_booking_permissions_manage_group = 'admin:sulten_booking_permissions_manage_group'
-admin__samfundet_booking_changelist = 'admin:sulten_booking_changelist'
-admin__samfundet_booking_add = 'admin:sulten_booking_add'
-admin__samfundet_booking_history = 'admin:sulten_booking_history'
-admin__samfundet_booking_delete = 'admin:sulten_booking_delete'
-admin__samfundet_booking_change = 'admin:sulten_booking_change'
+admin__samfundet_booking_permissions = 'admin:samfundet_booking_permissions'
+admin__samfundet_booking_permissions_manage_user = 'admin:samfundet_booking_permissions_manage_user'
+admin__samfundet_booking_permissions_manage_group = 'admin:samfundet_booking_permissions_manage_group'
+admin__samfundet_booking_changelist = 'admin:samfundet_booking_changelist'
+admin__samfundet_booking_add = 'admin:samfundet_booking_add'
+admin__samfundet_booking_history = 'admin:samfundet_booking_history'
+admin__samfundet_booking_delete = 'admin:samfundet_booking_delete'
+admin__samfundet_booking_change = 'admin:samfundet_booking_change'
 adminsamfundetbooking__objectId = ''
 admin__samfundet_closedperiod_permissions = 'admin:samfundet_closedperiod_permissions'
 admin__samfundet_closedperiod_permissions_manage_user = 'admin:samfundet_closedperiod_permissions_manage_user'
