+++ backend/samfundet/views.py
@@ -46,7 +46,6 @@ from .models.general import (
     Tag,
     User,
     Gang,
-    Venue,
     Image,
     Infobox,
     Profile,
@@ -60,15 +59,18 @@ from .models.general import (
     UserPreference,
     InformationPage,
 )
+from .models.venue import Booking, Table, Reservation, Venue
 from .serializers import (
     TagSerializer,
     GangSerializer,
     UserSerializer,
+    TableSerializer,
     ImageSerializer,
     EventSerializer,
     VenueSerializer,
     LoginSerializer,
     GroupSerializer,
+    BookingSerializer,
     InfoboxSerializer,
     ProfileSerializer,
     RegisterSerializer,
@@ -86,6 +88,7 @@ from .serializers import (
     UserPreferenceSerializer,
     InformationPageSerializer,
     OccupiedtimeslotSerializer,
+    ReservationCheckSerializer,
     UserForRecruitmentSerializer,
     RecruitmentPositionSerializer,
     RecruitmentAdmissionForGangSerializer,
@@ -671,4 +674,46 @@ class OccupiedtimeslotView(ListCreateAPIView):
             headers = self.get_success_headers(serializer.data)
             return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
 
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+# =============================== #
+#          Reservation            #
+# =============================== #
+
+
+class TableView(ModelViewSet):
+    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
+    serializer_class = TableSerializer
+    queryset = Table.objects.all()
+
+
+class BookingView(ModelViewSet):
+    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
+    serializer_class = BookingSerializer
+    queryset = Booking.objects.all()
+
+
+class ReservationCheckAvailabilityView(APIView):
+    permission_classes = [AllowAny]
+    serializer_class = ReservationCheckSerializer
+
+    def post(self, request: Request) -> Response:
+        serializer = self.serializer_class(data=request.data)
+        if serializer.is_valid():
+            if serializer.validated_data['reservation_date'] <= timezone.now().date():
+                return Response(
+                    {
+                        'error_nb': 'Reservasjoner må dessverre opprettes minst én dag i forveien.',
+                        'error_en': 'Unfortunately, reservations must be made at least one day in advance.'
+                    },
+                    status=status.HTTP_406_NOT_ACCEPTABLE
+                )
+            venue = self.request.query_params.get('venue', Venue.objects.get(slug='lyche').id)
+            available_tables = Reservation.fetch_available_times_for_date(
+                venue=venue,
+                seating=serializer.validated_data['guest_count'],
+                date=serializer.validated_data['reservation_date'],
+            )
+            return Response(available_tables, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
